requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
scikit-learn==1.3.2
nltk==3.8.1
plaid-python==11.0.0

